renamed to styles.js, added instructions for styles.js, order alphabetic
