+++ django-drf-react-quickstart/mcbcapp/frontend/src/views/LoginPage.js
@@ -4,7 +4,7 @@ import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import InputAdornment from "@material-ui/core/InputAdornment";
 import Grid from "@material-ui/core/Grid";
-import { loginPageStyles } from "../styling";
+import { loginPageStyles } from "../styles";
 import { withStyles } from "@material-ui/styles";
 
 class LoginPage extends React.Component {
