+++ django-drf-react-quickstart/mcbcapp/frontend/src/views/ProfilePage.js
@@ -3,7 +3,7 @@ import { makeStyles, useTheme, Typography } from "@material-ui/core";
 import MinistryIcon from "../components/MinistryIcon";
 import color from "@material-ui/core/colors/blueGrey";
 import { textAlign } from "@material-ui/system";
-import { profilePageStyles } from "../styling";
+import { profilePageStyles } from "../styles";
 import { withStyles } from "@material-ui/styles";
 
 class ProfilePage extends React.Component {
